+++ R/empirical.R
@@ -13,7 +13,12 @@
 #'
 #' @return Named list with elements:
 #' - `cov`: `matrix` empirical covariance.
-#' - `df_mat`: `matrix` to calculate Satterthwaite degree of freedom.
+#' - `g_mat`: `matrix` to calculate Satterthwaite degrees of freedom.
+#'
+#' @note
+#' This function used to return `df_mat`, which was equivalent to `crossproduct(g_mat)`. However,
+#' executing the cross product in C++ was a costly matrix multiplication, in particular when the number of coefficients
+#' and/or the number of subjects was large. Therefore this is now avoided and `g_mat` is returned instead.
 #'
 #' @keywords internal
 h_get_empirical <- function(tmb_data, theta, beta, beta_vcov, type) {
@@ -21,7 +26,24 @@ h_get_empirical <- function(tmb_data, theta, beta, beta_vcov, type) {
   assert_numeric(theta)
   n_beta <- ncol(tmb_data$x_matrix)
   assert_numeric(beta, finite = TRUE, any.missing = FALSE, len = n_beta)
-  assert_matrix(beta_vcov, mode = "numeric", any.missing = FALSE, nrows = n_beta, ncols = n_beta)
-  assert_subset(type, c("Empirical", "Empirical-Jackknife", "Empirical-Bias-Reduced"))
-  .Call(`_mmrm_get_empirical`, PACKAGE = "mmrm", tmb_data, theta, beta, beta_vcov, type)
+  assert_matrix(
+    beta_vcov,
+    mode = "numeric",
+    any.missing = FALSE,
+    nrows = n_beta,
+    ncols = n_beta
+  )
+  assert_subset(
+    type,
+    c("Empirical", "Empirical-Jackknife", "Empirical-Bias-Reduced")
+  )
+  .Call(
+    `_mmrm_get_empirical`,
+    PACKAGE = "mmrm",
+    tmb_data,
+    theta,
+    beta,
+    beta_vcov,
+    type
+  )
 }